Rename CDI examples test classes to avoid clashes in IDEs and polish
